requirements.txt
langchain==0.3.26
langchain-core>=0.3.66,<1.0.0
langchain-community==0.3.26
langchain-huggingface
langchain-pinecone==0.2.8
sentence-transformers
pypdf==5.6.1
fastapi
uvicorn
flask==3.1.1
sentence-transformers==4.1.0
python-dotenv==1.1.0
langchain-groq
python-multipart
-e .